Split Player to Player, PlayerComparator, PlayerActionType.
