+++ sources/Colonization/Backend/Player/PlayerComparator.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include <Colonization/Backend/Player/Player.hpp>
+
+namespace Colonization
+{
+class Player;
+typedef bool (*PlayerComparator) (const Player *first, const Player *second);
+namespace PlayerComparators
+{
+bool HigherPoints (const Player *first, const Player *second);
+}
+}
