+++ sources/Colonization/Backend/VictoryProgressUpdater.cpp
@@ -17,8 +17,9 @@
 #include <Colonization/Backend/PlayersManager.hpp>
 #include <Colonization/Backend/UnitsManager.hpp>
 #include <Colonization/Backend/TradeProcessor.hpp>
-#include <Colonization/Utils/Network/NetworkUpdateCounter.hpp>
+#include <Colonization/Backend/Player/PlayerComparator.hpp>
 
+#include <Colonization/Utils/Network/NetworkUpdateCounter.hpp>
 #include <Colonization/Utils/Serialization/Categories.hpp>
 #include <Colonization/Utils/Serialization/AttributeMacro.hpp>
